requirements.txt
beautifulsoup4==4.5.3
certifi==2016.9.26
future==0.16.0
lxml==3.7.2
pymongo==3.4.0
python-telegram-bot==5.3.0
requests==2.12.5
urllib3==1.20
